feat: auto-deploy slash commands on startup and remove CLIENT_ID requirement
Commands are now automatically registered via Discord REST API when the bot
starts, removing the need for manual `npm run deploy` and the CLIENT_ID
environment variable. Includes duplicate command detection and uses the
project's structured logger.

Closes #116
