+++ .env.example
@@ -1,7 +1,6 @@
 # Discord Bot Configuration
-# Get these from https://discord.com/developers/applications
+# Get your token from https://discord.com/developers/applications
 TOKEN=your_discord_bot_token_here
-CLIENT_ID=your_discord_client_id_here
 
 # Optional: Spotify Configuration
 # To enable Spotify support, set SPOTIFY_ENABLED=true and provide your credentials
