+++ docs/index.html
@@ -502,9 +502,6 @@
                                 <div class="config-line">
                                     <span class="config-key">TOKEN</span>=<span class="config-value">your_discord_bot_token</span>
                                 </div>
-                                <div class="config-line">
-                                    <span class="config-key">CLIENT_ID</span>=<span class="config-value">your_client_id</span>
-                                </div>
                                 <div class="config-line">&nbsp;</div>
                                 <div class="config-line">
                                     <span class="config-comment"># Lavalink &mdash; optional, leave empty for free public servers</span>
@@ -529,32 +526,6 @@
 
                 <div class="step">
                     <div class="step-number">03</div>
-                    <div class="step-content">
-                        <h3 class="step-title">Deploy Commands</h3>
-                        <div class="terminal">
-                            <div class="terminal-header">
-                                <span class="terminal-dot"></span>
-                                <span class="terminal-dot"></span>
-                                <span class="terminal-dot"></span>
-                                <span class="terminal-title">Terminal</span>
-                            </div>
-                            <div class="terminal-body">
-                                <div class="terminal-line">
-                                    <span class="prompt">$</span>
-                                    <span class="command">docker compose run --rm bot npm run deploy</span>
-                                </div>
-                                <div class="terminal-output"><i class="bi bi-check-lg"></i> Successfully registered 12 commands</div>
-                            </div>
-                            <button class="copy-btn" data-copy="docker compose run --rm bot npm run deploy">
-                                <span class="copy-icon"><i class="bi bi-clipboard"></i></span>
-                                <span class="copy-text">Copy</span>
-                            </button>
-                        </div>
-                    </div>
-                </div>
-
-                <div class="step">
-                    <div class="step-number">04</div>
                     <div class="step-content">
                         <h3 class="step-title">Launch BeatDock</h3>
                         <div class="terminal">
@@ -717,10 +688,6 @@
                         <div class="config-row">
                             <code class="config-var">TOKEN</code>
                             <span class="config-desc">Discord bot token from Developer Portal</span>
-                        </div>
-                        <div class="config-row">
-                            <code class="config-var">CLIENT_ID</code>
-                            <span class="config-desc">Discord application/client ID</span>
                         </div>
                     </div>
                 </div>
