+++ src/events/ready.js
@@ -1,21 +1,22 @@
-const { ActivityType } = require('discord.js');
+const { Events } = require('discord.js');
 const { generateInviteUrl } = require('../utils/inviteUrl');
+const deployCommands = require('../utils/commandDeployer');
 const logger = require('../utils/logger');
 
 module.exports = {
-    name: 'ready',
+    name: Events.ClientReady,
     once: true,
     async execute(client) {
         logger.info(`Bot ready as ${client.user.tag}`);
 
-        // Initialize Lavalink
+        await deployCommands(client);
+
         await client.lavalink.init({ ...client.user });
         logger.info('Lavalink initialized');
 
         const inviteUrl = generateInviteUrl(client.user.id);
         logger.info(`Invite URL: ${inviteUrl}`);
 
-        // No initial presence - will be set when music starts playing
         client.user.setActivity(null);
     },
-}; 
+};
